Don't render back button label just yet
We've landed it, but the PR that implemented truncation of too long titles was reverted because it wasn't tested on Android. Since @satya164 is about to release a new version, please consider merging that one so that people don't end up having issues with back button being too long.

The fix I am working on is to land tomorrow latest, but this is in case one wants to release something in the meantime.