Send destructured screenProps to Component
Hi!

I have used NavigationExperimeltal a lot with naviagation in Redux tree. I just started to migrate to this awesome library but one show stopper was this. I really dont understand why you are hiding the props sent down from my reducer function in an object called screenProps. My goal is to be able to navigate to normal components. Now when the props are inside the screenProps object i need to split up components into "go To"-able components and normal components => probably not what was intended.

Anyways i use a HOC function for now to solve this, let me know what you guys think and if i missed you something. Ill continue to test the lib if you guys get back to me with some small feedback. 

Let me know if you are interested in a minimalistic repo with this issue explained.