add 2nd params argument to StackRouter.getActionForPathAndParams
It still falls back to parsing the query string if no params argument is provided, but now TabRouter.getActionForPathAndParams which might call StackRouter.getActionForPathAndParams with a 2nd params argument will work correctly. 

In addition, if called directly (as may be the case in apps with custom deep-linking needs [which is my primary case]), it will function correctly without users having to build a query string. 

This also solves the problem that deserialization of query strings is currently implemented without support for cases like nested arrays within the query string.