fix($TabRouter): make `paths` option work
The `paths` option to TabNavigator and TabRouter was not in fact implemented, yet it was documented. Here's the small fix for it.

That said, I'm not even sure what the purpose of the `paths` option is, as there's nothing to override since the `routeConfigs` are defined in the exact same place, rather than imported and ***reused***. Since there is no *re-use* capability currently for paths, there is no need for a `paths` option. 

See my issue related to this: https://github.com/react-community/react-navigation/issues/1665