passActiveRouteOption to support nested configs
New API to support nested configuration, such as the header title within the “StacksOverTabs” example. In v2, this nested configuration behavior is explicit.

While I’m happy with the new API and don’t see any bugs, this is a pretty invasive change to the core code, and we should review the changes to the navigation prop. The following new pieces of API have been added:

- passActiveRouteOption. See how its used in `StacksOverTabs.js`
- router.childRouters, a map of routeName to child router. Needed for static access to child router
- navigation.router refers to the router that maintains the navigation state. This should be safe
- navigation.childrenNavigation is a map of route key to child navigation objects. This might be ok, but might be a footgun
- navigation.getScreenProps, added as a convenience and I think it is safe. Maybe should rename to discourage screenProps usage altogether.

Lots of changed code:

- dangerously get parent seems less dangerous now, so I use it within passActiveRouteOption
- subscriptions (addListener) are now cached inside the child navigation creator

So this does seem to address the problem fine, but I think we should be careful moving forward here. Careful code review would be appreciated!