Fix bug in DrawerView
When attempting to navigate to a drawer item while the drawer's state
is 'opening' or 'closing' results in a state where the underlying navigator's
screen correctly changes but the drawer fails to automatically dismiss itself.

Instead, we should allow the drawer to respond to updates and change it's state
even if it's in the middle of 'opening' or 'closing'.