fix: focus/blur events emitted with slight delay
The lifecycle ‘focus’ and ‘blur’ events are called with a slight delay and in next animation-frame on react-native; and after the DOM is updated on web.

This causes problems with applications which want to use the underlying animated-value upon receiving focus of a screen, such as shared-element transtions.

This fix solves that by running the event-emitter code directly after the mutations have been applied using `useLayoutEffect`.