Always set the isOpen when the drawer is toggled
This fixes an issue for us in our app for multi-tasking on iPad. When we were on iPad, if we closed the drawer and went narrow (to mobile size) the drawer was still open and would never close. I am not sure why isOpen was always true when we called `toggleDrawer(false)`. This might not be a good/valid fix but we didn't have issues after testing this.