fix(useScrollToTop): handling multiple tab bars
Fixes #8586
If a screen is inside a bottom tab bar and a top tab bar, only a click on the first parent (top tab bar) will scroll to top. Clicking on the bottom tab bar won't.
With this fix, all parents being a tab bar add an event listener to scroll to top. It will only scroll to top when the current view is inside one of the tab bars.