fix: warn for components starting with lower case names
While passing a component with an uncapitalized name works, it can mean that the user is trying to pass a regular function and not a component. Regular functions are usually defined inside components causing state to be lost on re-render if used as components. This warning is supposed to catch those cases.