fix: improve when back button is shown in nested native stacks
Currently when a Native Stack is nested in another Stack, it doesn't respect the
parent navigation's back stack. This change makes a few adjustments so that if
the parent has a back stack:

- If the user specifies a custom back button, the `canGoBack` and `label`
parameters will now receive the proper data so they can render a back button in
nested stack.
- If JS stack is nested inside a Native stack, it'll now respect the parent's
back stack and show a back button.
- On Web, Native stack nested inside or Native or JS stack will show a back
button.
- On Android, Native Stack nested inside Native stack will show a back button -
since the back button is handled natively, we can't make it show when Native
Stack is nested inside JS Stack
- On iOS, JS stack nested inside Native stack will show a back button - since the
back button is handled natively, we can't make it show when Native Stack is
nested inside JS or Native Stack (see
https://github.com/software-mansion/react-native-screens/issues/1460)

starting # with '#' will be ignored, and an empty message aborts the commit.