fix: restore native behavior of back button
Unfortunately the changes from #10761 caused an issue I reported on the
react-native-screens repo. (See software-mansion/react-native-screens#1589)

Essentially, the documentation states that per native behavior, the back button
on iOS will automatically use the title of the previous screen or "Back" if
there's not enough space. However, with the latest version of react-navigation,
the back button was the title of the previous screen no matter what. In some
cases, this caused the title of the current screen to be completely pushed off
the screen.

In UIKit this happens when the title of the back button is deliberately set, or
when a custom UIBarButtonItem is set as the back button. `react-native-screens`
creates a custom UIBarButtonItem when the header config component has a value
defined for `headerBackTitle`. So by leaving headerBackTitle undefined here, we
let the system automatically determine what to set as the back button title. In
my testing, system will choose the correct value even if the previous screen's
`title` or `headerTitle` is set differently from the route name.