fix: add additional check before running `animate` in `componentDidUpdate` (#10871)
Changes introduced in #10767 created a regression reported in #10856 which made it impossible to close a modal with a gesture.

As we want to change pointer-events on Card in componentDidUpdate previous use of setNativeProps in this context made a lot of sense. Since setNativeProps is going to be removed in the new architecture we had to migrate it to the component state as stated in the migration guide.

This PR adds an additional `closing !== prevProps.closing` check on top of #10767 within componentDidUpdate in Card.tsx.