refactor!: drop independent prop in favor of a wrapper component (#10907)
This drops the `independent` prop from the container in favor of a component. In case of cases such as mini-apps, this allows the parent app to disconnect the child navigation tree from itself.

Another reason is that a lot of newbies added `independent` prop to avoid error and got confused when nested navigation didn't work. By moving it into a component and making it harder to do it accidentally, maybe this can be avoided.