fix: fix drawer toggle race condition (#11036)
After navigating programmatically, the drawer is toggling unexpectedly if the JS thread is busy due to a race condition between the drawer animation and navigation state. Under extreme cases, the drawer may open and close multiple times before stopping.

This is because after the toggle animation is complete, the drawer always dispatches a navigation action. This is redundant in the case where the drawer opening or closing was done programmatically, and typically this should not cause another animation since the navigation state usually will not have changed. However, if the navigation state has already changed, the drawer may re-open (or close) and trigger another toggle animation.

This PR fixes this issue by only dispatching a navigation action when the drawer was toggled by the user and not programmatically.