feat: add option to show tabs on the side (#11578)
**Motivation**

This adds ability to render the tab bar on left/right as a sidebar. It
useful on larger screens where we may want to take advantage of the
available horizontal screen and show a sidebar instead of a bottom tab
bar.

**Test plan**

Open the "Bottom Tabs" example and resize the screen to see the
different modes of the tab bar. Tap on the "<<" button on the bottom
left to switch between a smaller and larger sidebar.

**Video**


https://github.com/react-navigation/react-navigation/assets/1174278/224beaa9-8e26-4d94-a70d-11e54c37d944