feat: add a layout prop for navigators (#11614)
**Motivation**

This adds a new `layout` prop to navigators. It can be useful for
augmenting the navigators with additional UI with a wrapper. The
difference from adding a regular wrapper is that the code in `layout`
callback has access to the navigator's state, options etc.

**Test plan**

Tested by implementing a breadcrumb UI in stack navigator in the example
app.


https://github.com/react-navigation/react-navigation/assets/1174278/5ee2b7f5-f000-4d43-a147-33a1e1562b4c