feat: add a button element to elements package
This adds a simple `Button` component to make it easier for quick demos, examples and learning material etc.

It implements part of the material design guidelines: https://m3.material.io/components/buttons/overview

There are 3 variants implemented:

- `'plain'`
- `'tinted'`
- '`filled'`

These variants also exist in iOS design guidelines, so it makes the buttons somewhat platform agnostic. Though exact measurements will slightly differ.

Since the `Button` is part of React Navigation, it has built-in support for navigating to screens, and renders an anchor tag on the Web when used for navigation:

```js
<Button screen="Profile" params={{ userId: 'jane' }}>
  View Jane's Profile
<Button>
```

It can also be used as a regular button:

```js
<Button onPress={() => { /* do something */ }}>
  Do something
</Button>
```