feat: add layout and screenLayout props for screens
This adds a new `layout` prop that can be used to wrap screens.

It makes it easier to provide things such as a global error boundary and suspense fallback for a group of screens without having to manually add HOCs for every screen separately.

See https://github.com/react-navigation/react-navigation/discussions/11152