feat: preloading for stack navigator (#11733)
This is yet another follow-up of #11702. Now, it is about implementing
preloading for the stack navigator.

The first essential part is merged from #11727 and is about creating
temporary descriptors for routes which are not yet in the navigation
state, but only preloaded.

The remaining part is adjusting the logic for the preloading flow.
Particularly, a lot of changes were needed for handling cards, where
`gesture` is `undefined` and making the animation run a later for those
routes (not on the initial mount, but on navigating). The remaining
changes are not critical.

Precise comments are included in the comments.

Please review the tests and the video below.


https://github.com/react-navigation/react-navigation/assets/25709300/830a7dfc-cf6a-4117-b38d-7f6326934af1

---------

Co-authored-by: Satyajit Sahoo <satyajit.happy@gmail.com>