fix: handle conflicting nested path params for linking (#11849)
**Motivation**
When a path configuration contains multiple path parameters with the
same name, the last path parameter value wins when constructing the
navigation state.

For example consider the following scenario:

```ts
const path = '/foo/42/bar/43';

  const config = {
    initialRouteName: 'Foo',
    screens: {
      Foo: {
        path: 'foo/:id',
        screens: {
          Bar: {
            path: 'bar/:id',
          },
        },
      },
    },
  };

  const state = {
    routes: [
      {
        name: 'Foo',
        params: { id: '43' }, // PROBLEM: This should be 42
        state: {
          routes: [
            {
              name: 'Bar',
              params: { id: '43' },
              path,
            },
          ],
        },
      },
    ],
  }
```

This PR updates the resolution logic to assign positions to each of the
path segments, so that we can assign the correct values to the screen in
the navigation state based on the position which it appears in the path,
and not just the name.

**Test plan**
* Copy the test in this PR, checkout to main and run the tests
* You will notice that the test fails to resolve the right params to the
route `'Foo'`
* Now checkout to this PR's branch and run all tests and notice that
they pass

**Considerations**
* Another approach could be to include validation logic that prevents
this scenario from happening by validating that there are no repeated
path params with the same name.
* This however, could be a breaking change, and might call for
additional functionality as "aliasing" to be able to name path params in
the config, and then translate them to a different name before being
passed to the screen.
* Alternatively, in the implementing screen, the developer could update
the route param name to match what is in the new path parameter
* OR, in the implementing screen, accept either the original param name
or deep link param name and pick whichever one is available

Happy to adjust this PR to the validation approach if desired, or close
in favor of the user-land mitigations. Thanks!

---------

Co-authored-by: Satyajit Sahoo <satyajit.happy@gmail.com>