fix: only replace trailing slash on decode (#11899)
**Motivation**

I first noticed this in `@react-navigation/core@6.4.14`. I think it
happened in
https://github.com/react-navigation/react-navigation/pull/11869

When you decode a path param that has an encoded `/` it removes the
encoded param and leaves the trailing param in state.

**Test plan**
Create a route that have a path param with a`%2F` in it, verify that it
comes out correctly decoded