test: change queryByText to getByText in useUnhandledLinking.test.tsx (#11913)
**Motivation**

In `useUnhandledLinking.test.tsx` file, we're trying to query the
element with given label and press it with `fireEvent.press`.
Unfortunately, in my case Typescript prints several warnings about the
wrong return type of passed argument. In fact, Typescript is correct
there - we're trying to pass the argument of type `ReactTestInstance |
null` to the function that accepts arguments of type
`ReactTestInstance`. Because of that, I'm willing to change
`queryByText` to `getByText` in places where the tests are trying to
press the element with queried label.

I'm also leaving remains of the `queryByText` in the expect functions
intentionally, since we're still expecting that some elements will be
null there.

**Test plan**

Passing CI should be enough to check if that works! ✅