refactor: drop unmountOnBlur in favor of popToTopOnBlur (#11896)
This drops the `unmountOnBlur` option of bottom tabs and drawer in favor
of a `popToTopOnBlur` option.

In many cases, the requirement is to return to the first screen of the
stack nested in tabs. However using `unmountOnBlur` in those scenarios
would also destroy any state of the screen. It's also slower to remount
a nested navigator on tab navigation.

The `popToTopOnBlur` option provides an alternative approach - it pops
the screens on a nested stack to go back to the first screen in the
stack. It's much faster than remounting the nested navigator, and also
doesn't destroy the state of the first screen in the stack.

It's still possible to achieve the old behavior of `unmountOnBlur` by
using the `useIsFocused` hook in the screen:

```js
const isFocused = useIsFocused();

if (!isFocused) {
  return null;
}
```