fix: fix back icon not showing in iOS production builds in Expo, fixes #11792 (#12019)
**Motivation**

I've been seeing the issue described by
https://github.com/react-navigation/react-navigation/issues/11792 where
the back button is not shown on iOS. It's a frustrating problem since
the issue – back buttons on a Stack Navigator not showing on iOS – only
happens in production builds.

**Test plan**

I ported the React Native code for `renderBackImage` in
`HeaderBackButton` along with all assets and used that as a starting
point to troubleshoot the issue. Removing the `1.5x` variants fixes the
problem for me. I'm enough of a React Native dunce that I don't totally
understand why those two files cause issues, but removing those files
seems fine since that scale doesn't seem to be supported anyway
according to [the Image docs](https://reactnative.dev/docs/images).

> Note - it could also be the case that the `1x` variants are
superfluous and should be removed, but I'll leave that to the
maintainers to decide

I'd submit a test plan but as the others have reported in the issue
linked above, this is a production-only bug for me.

Co-authored-by: Satyajit Sahoo <satyajit.happy@gmail.com>