test: add maestro tests with playwright adapter (#12077)
This adds some basic Maestro tests. Note that most tests are not testing much, but once we have basic setup, we can update them as we go.

Currently need to be run locally as I haven't been successful with CI setup for Android yet.

This also adds custom logic that runs the Maestro tests to run with Playwright for Web, so we can reuse the same tests. We'll need to update the logic as we add more tests.