refactor: Export NativeStackHeaderRightProps and NativeStackHeaderLeftProps (#12188)
**Motivation**

I was building a custom header right component, and noticed the props
are typed as NativeStackHeaderRightProps, but I couldn't import that
type from anywhere. Of course it was trivial to just make my own type,
but it would be nice to be able to just import the type from the package
since it's actually available in the types, just not exported from the
barrel file. I also am exporting NativeStackHeaderLeftProps for the same
reasons.

**Test plan**

It's just exporting a type from the barrel file, so not quite sure
if/how it could be tested. No actual types have been changed.

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>