refactor: move screen specific logic for native stack to own component (#12201)
Currently there is a lot of logic to make `Screens` work properly in
native-stack, including:

- Making sure screens is enabled and mark it for native-stack
- Making sure screens knows that it contains a large title
- Rendering a `DebugContainer` for `LogBox`
- Making sure the `DebugContainer` contains appropriate logic and
wrapper
- Rendering a nested stack to be able to show header in modals

This diff consolidates this logic to a new component
`ScreenStackContent`. This makes it possible to move the above logic to
the `react-native-screens` package by moving this component. Moving the
logic will make it easier to fix bugs in screens.

**Next Steps**: Once we merge this, we should move the `DebugContainer`
and `ScreenStackContent` to `react-native-screens` package and update
native-stack's code to use the exported `ScreenStackContent` from
`react-native-screens`.