feat: provide gesture in DrawerGestureContext (#12326)
**Motivation**

The drawer gesture can conflict with other gestures in the app. To
coordinate multiple gestures, the app code can use APIs like
`requireExternalGestureToFail`. However, since the gesture is hidden in
the library, this is not currently possible without patching. This
exposes the gesture to the app code so you can use
`requireExternalGestureToFail` with it.

```js
import { DrawerGestureContext } from 'react-native-drawer-layout';

// ...
const drawerGesture = useContext(DrawerGestureContext)
const nativeGesture = Gesture.Native()
  .requireExternalGestureToFail(drawerGesture)

// ...

<GestureDetector gesture={nativeGesture}>
  <PagerView
    // ...
  />
</GestureDetector>
```

Note that the existing `configureGestureHandler` is not enough because
we need it in the other direction: *other* gestures (deeper in the app)
need access to it. It would not be possible to capture it in a closure
because by the point `configureGestureHandler` runs, the owner component
has already executed and passed the data down.

<s>I opted to use a render prop instead of actually creating a context
in order to keep this addition zero-cost. If this was a context exposed
by the library, we'd have to either memoize the `pan` gesture (to avoid
it getting redefined on drawer re-renders) or we'd have to pay the price
of React doing a context search traversal every time it changes. Seems
unfortunate if the majority of this library's users don't care about the
gesture. Render prop lets the app decide whether it cares.</s>

Re-done as context as requested in review. Added some memoization to
avoid it changing on every render.

**Test plan**

Verify you're able to access `gesture` on mobile. On web, it should be
`undefined`.

Verify you can use it with RNGH APIs as usual. For example,
https://github.com/bluesky-social/social-app/pull/7007 fixes a
longstanding problem with the Bluesky app — the drawer swipe barely
worked on iOS. The reason it has been broken for the last two years is
because this library did not expose the gesture, and we did not know
that it needs to be coordinated.