fix: nanoid vulberable version (#12328)
This PR is to update the version of the **nanoid** library used in core,
native and routers packages.
**Motivation**
There were 9 vulnerabilities with **low** severity reported in
**nanoid** library of version 3.3.7.
The next closest version of the library with no vulnerabilities is
3.3.8.
Link to the vulnerability report:
https://github.com/advisories/GHSA-mwcw-c2x4-8c55

This migration is already tested in a side project react native app. No
difference observed, works fine. This migration solved all
vulnerabilities.

---------

Co-authored-by: Khushil Sindhwad <67272210+khushilms-cs19@users.noreply.github.com>