fix: flush state updates before `setPreventRemove` called (#12335)
**Motivation**

In an app I've been developing, navigating to a screen that contained a
call to `usePreventRemove(true, ...)` would throw this error:

https://github.com/react-navigation/react-navigation/blob/2a745c8c598f95fcec5bbf5442045478d4046663/packages/core/src/PreventRemoveProvider.tsx#L69-L71

This only started happening in v7, not v6. I'm not sure if it only
happens when navigating to a child navigator, but the issue seems to
stem from changes to when scheduled updates to navigation state are
executed here:

https://github.com/react-navigation/react-navigation/blob/2a745c8c598f95fcec5bbf5442045478d4046663/packages/core/src/useScheduleUpdate.tsx#L11-L21

In v6, `scheduleUpdate` would have executed `callback` synchronously and
the callback to update the state would be executed immediately, before
child navigators would be (re)rendered.


https://github.com/react-navigation/react-navigation/blob/fd2dda1bc43804f1a4c1cb9ee2739b39e8163210/packages/core/src/useSyncState.tsx#L43-L51

In v7 however, `scheduleUpdate` doesn't execute `callback` immediately
and so the navigation state would be out of sync with latest state in
`PreventRemoveProvider` until next render cycle in turn throwing an
error.

https://github.com/react-navigation/react-navigation/blob/2a745c8c598f95fcec5bbf5442045478d4046663/packages/core/src/PreventRemoveProvider.tsx#L65-L67

**Test plan**

By flushing scheduledUpdates in `useLayoutEffect` instead of `useEffect`
it seems to ensure that the latest navigation state is propagated to
`PreventRemoveProvider` as expected and in turn the call to
`setPreventRemove` will not throw an error when called in `useEffect`
hook.

Of course, I cannot quite foresee how this might affect other parts of
the code.